Various new implementations

+++ src/utils.py
@@ -1,14 +1,18 @@
 import configparser
+import os
+
 
 def loadconfig():
     config = configparser.ConfigParser()
-    config['Mullvad'] = {'MullvadAccount': '123456789'}
-    config['TelegramBot'] = {'Token': 'https://t.me/token/api', 'PoolInterval': '30', 'Enabled': 'false'}
 
-    try:
-        configfile = 'config.ini'
-        config.read(configfile)
-    except:
+    if not os.path.exists('config.ini'):
         with open('config.ini', 'w') as configfile:
+            config['Mullvad'] = {'MullvadAccount': '123456789'}
+            config['TelegramBot'] = {'Token': 'https://t.me/token/api', 'PoolInterval': '30', 'Enabled': 'false'}
             config.write(configfile)
-    return config
+            print('write')
+            return config
+    else:
+        config.read('config.ini')
+        print('read')
+        return config
